fixed segfaults when runing into walls

+++ Hero.h
@@ -35,4 +35,4 @@ Hero xteleport(Hero, const Map);
 
 Hero xsustain(Hero, const Map, const Input);
 
-Ray xcalc(const Hero hero, const Hit, const float level, const int yres);
+Ray xcalc(const Hero hero, const Hit, const int level, const int yres);
